requirements.txt
Telethon
python-dotenv
pybit
colorama
requests
websocket-client